Revert "Feature auth 14 implement email services in user services"
