--- api/models/base_model.py
@@ -1,14 +1,0 @@
-import uuid
-
-from django.db import models
-
-
-class ECOMBaseModel(models.Model):
-    id = models.UUIDField(
-        default=uuid.uuid4, unique=True, primary_key=True, null=False, editable=False
-    )
-    created_at = models.DateTimeField(auto_now_add=True, editable=False)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        abstract = True
