+++ api/services/encryption_service.py
@@ -12,16 +12,14 @@ class EncryptionServices:
             f"{os.environ.get('ENCRYPTION_PASSWORD'):<32}".encode("utf-8")
         )
         encryptor = Fernet(key=key)
-        password = base64.b64encode(password.encode("utf-8"))
-        encrypted_password = encryptor.encrypt(password)
+        encrypted_password = encryptor.encrypt(password.encode("utf-8"))
         return encrypted_password
 
     def decrypt(self, encrypted_password):
         load_dotenv()
         key = base64.b64encode(
-            f"{os.environ.get('ENCRYPTION_PASSWORD'):<32}".encode("utf-8")
+            f"{os.environ.get('ENCRYPTION_PASSWORD')}".encode("utf-8")
         )
         encryptor = Fernet(key=key)
         decrypted_password = encryptor.decrypt(encrypted_password).decode("utf-8")
-        decrypted_password = base64.b64decode(decrypted_password).decode("utf-8")
         return decrypted_password
