+++ api/tests/views/test_create_users_view.py
@@ -1,12 +1,9 @@
 import pytest
-from rest_framework import serializers
-from rest_framework.test import APIClient
-
-from api.serializers.ecom_user_serializer import ECOMUserSerializer
+from rest_framework.test import APITestCase
 
 
 @pytest.mark.django_db
-class TestCreateUsersView:
+class TestCreateUsersView(APITestCase):
     def test_create_user(self):
         data = {
             "username": "koushikmallik",
@@ -16,10 +13,8 @@ class TestCreateUsersView:
             "password1": "1234567",
             "password2": "1234567",
         }
-
-        client = APIClient()
         url = "http://localhost/api/v2/create-users"
-        response = client.post(url, data, format="json")
+        response = self.client.post(url, data, format="json")
         assert response
         response = response.json()
         if response.get("error"):
@@ -28,63 +23,3 @@ class TestCreateUsersView:
             assert response.get("token")
             assert response.get("token").get("refresh")
             assert response.get("token").get("access")
-
-    @pytest.mark.parametrize(
-        "username,email,fname,lname,password1,password2",
-        [
-            (
-                "koushikmallik",
-                "abcdef@google.com",
-                "Koushik123",
-                "Google",
-                "1234567",
-                "1234567",
-            ),
-            (
-                "koushikmallik",
-                "abcdefgooglecom",
-                "Koushik",
-                "Google",
-                "1234567",
-                "1234567",
-            ),
-            (
-                "koushikmallik",
-                "abcdef@google.com",
-                "Koushik",
-                "Google123",
-                "1234567",
-                "1234567",
-            ),
-            ("koushikmallik", "abcdef@google.com", "Koushik", "Google", "1234", "1234"),
-            (
-                "koushikmallik",
-                "abcdef@google.com",
-                "Koushik",
-                "Google",
-                "123456",
-                "123467",
-            ),
-            ("kou", "abcdef@google.com", "Koushik", "Google", "123456", "123456"),
-        ],
-    )
-    def test_create_user_negative(
-        self,
-        username: str,
-        email: str,
-        fname: str,
-        lname: str,
-        password1: str,
-        password2: str,
-    ):
-        data = {
-            "username": username,
-            "email": email,
-            "fname": fname,
-            "lname": lname,
-            "password1": password1,
-            "password2": password2,
-        }
-        with pytest.raises(serializers.ValidationError):
-            serializer = ECOMUserSerializer()
-            serializer.create(data)
